LatestViewController추가
