+++ src/main/java/com/idle/shoppingmall/RequestDTO/Payment/PaymentAddRequest.java
@@ -1,4 +1,4 @@
-package com.idle.shoppingmall.RequestDTO;
+package com.idle.shoppingmall.RequestDTO.Payment;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
