+++ src/main/java/com/idle/shoppingmall/RequestDTO/Payment/PaymentDeleteRequest.java
@@ -1,4 +1,4 @@
-package com.idle.shoppingmall.RequestDTO.Product.Delete;
+package com.idle.shoppingmall.RequestDTO.Payment;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
