+++ src/main/java/com/idle/shoppingmall/RequestDTO/Product/Search/ProductSearchRequest.java
@@ -1,0 +1,12 @@
+package com.idle.shoppingmall.RequestDTO.Product.Search;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@AllArgsConstructor
+@Getter
+@NoArgsConstructor
+public class ProductSearchRequest {
+    private String pd_name;
+}
