+++ src/main/java/com/idle/shoppingmall/ResponseDTO/Product/ProductListResponse.java
@@ -18,13 +18,11 @@ public class ProductListResponse extends CommonResponse {
     private LocalDateTime created_at;
     private int count_love;
 
-    public ProductListResponse(int code, String msg, Long product_id, String pd_name, int pd_price, String brand_name, String pd_category, Long created_who, LocalDateTime created_at, int count_love) {
+    public ProductListResponse(int code, String msg, Long product_id, String pd_name, int pd_price, String pd_category, Long created_who, LocalDateTime created_at, int count_love) {
         super(code, msg);
         this.product_id = product_id;
         this.pd_name = pd_name;
         this.pd_price = pd_price;
-
-        this.brand_name = brand_name;
         this.pd_category = pd_category;
         this.created_who = created_who;
         this.created_at = created_at;
