+++ src/main/resources/application.properties
@@ -1,8 +1,8 @@
 spring.application.name=ShoppingMall
 
-spring.datasource.url=jdbc:mysql://db-lpeqn-kr.vpc-pub-cdb.ntruss.com:3306/shopping_mall?serverTimezone=Asia/Seoul
-spring.datasource.username=mini_ddl
-spring.datasource.password=bitcamp123!@#
+spring.datasource.url=jdbc:mysql://db-lsa9l-kr.vpc-pub-cdb.ntruss.com:3306/Mini_Shoppingmall?serverTimezone=UTC&useSSL=false&allowPublicKeyRetrieval=true
+spring.datasource.username=mini
+spring.datasource.password=1q2w3e4r!
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 
 # mybatis mapper ????
